chore(demo): new expandable row data table example
