+++ projects/demo/src/modules/components/table/examples/10/index.less
@@ -1,0 +1,41 @@
+th {
+    white-space: nowrap;
+}
+
+.table {
+    inline-size: 36rem;
+}
+
+.chips {
+    display: flex;
+    gap: 0.5rem;
+    flex-wrap: wrap;
+    padding: 0.5rem;
+}
+
+.last-name-col {
+    inline-size: 10rem;
+}
+
+.expand-heading-row {
+    cursor: pointer;
+
+    td {
+        background: var(--tui-background-neutral-1);
+    }
+
+    &:hover td {
+        background: var(--tui-background-neutral-1-hover);
+    }
+}
+
+.expand-heading-row-custom {
+    td {
+        background: var(--tui-background-neutral-1);
+    }
+
+    td.expand-heading-row-custom__expand:hover {
+        cursor: pointer;
+        background: var(--tui-background-neutral-1-hover);
+    }
+}
