+++ projects/demo/src/modules/components/table/examples/11/index.html
@@ -1,0 +1,52 @@
+<table
+    tuiTable
+    [columns]="columns"
+    [direction]="direction()"
+    [tuiSortBy]="sortBy()"
+    (tuiSortChange)="sortChange($event)"
+>
+    <thead>
+        <tr tuiThGroup>
+            <th
+                *tuiHead="'name'"
+                tuiSortable
+                tuiTh
+                [maxWidth]="500"
+                [minWidth]="200"
+                [resizable]="true"
+            >
+                Name
+            </th>
+            <th
+                *tuiHead="'id'"
+                tuiSortable
+                tuiTh
+            >
+                ID
+            </th>
+            <th
+                *tuiHead="'color'"
+                tuiSortable
+                tuiTh
+            >
+                Color
+            </th>
+        </tr>
+    </thead>
+    <tbody
+        tuiTbody
+        [data]="data()"
+    >
+        <tr
+            *ngFor="let item of data()"
+            tuiTr
+        >
+            <td
+                *tuiCell="'id'"
+                tuiTd
+            >
+                {{ item.id }}
+            </td>
+        </tr>
+    </tbody>
+</table>
