+++ projects/demo/src/modules/components/table/examples/9/index.less
@@ -1,41 +1,43 @@
-th {
-    white-space: nowrap;
-}
+@import '@taiga-ui/core/styles/taiga-ui-local';
 
 .table {
-    inline-size: 36rem;
-}
+    tbody {
+        tr {
+            td {
+                .transition(background);
+            }
 
-.chips {
-    display: flex;
-    gap: 0.5rem;
-    flex-wrap: wrap;
-    padding: 0.5rem;
-}
+            &.child td {
+                background: var(--tui-background-base-alt);
+            }
 
-.last-name-col {
-    inline-size: 10rem;
-}
+            &:hover {
+                td {
+                    background: var(--tui-background-base-alt);
+                }
 
-.expand-heading-row {
-    cursor: pointer;
+                &.child td {
+                    background: var(--tui-background-neutral-1-hover);
+                }
+            }
+        }
 
-    td {
-        background: var(--tui-background-neutral-1);
+        tui-table-expand {
+            animation-duration: 0.5s !important;
+        }
     }
 
-    &:hover td {
-        background: var(--tui-background-neutral-1-hover);
-    }
-}
-
-.expand-heading-row-custom {
-    td {
-        background: var(--tui-background-neutral-1);
+    [tuiTh] {
+        border-top: none;
     }
 
-    td.expand-heading-row-custom__expand:hover {
-        cursor: pointer;
-        background: var(--tui-background-neutral-1-hover);
+    [tuiTh],
+    [tuiTd] {
+        border-inline-start: none;
+        border-inline-end: none;
+    }
+
+    button[disabled].chevron-btn::after {
+        cursor: initial;
     }
 }
