+++ projects/demo/src/modules/components/table/examples/9/index.ts
@@ -1,34 +1,87 @@
-import {AsyncPipe, NgForOf} from '@angular/common';
+import {NgForOf, NgIf} from '@angular/common';
 import {Component} from '@angular/core';
+import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
-import {TuiTable, TuiTableExpand} from '@taiga-ui/addon-table';
-import {TuiMapperPipe} from '@taiga-ui/cdk';
-import {TuiButton, TuiFormatNumberPipe, TuiHint} from '@taiga-ui/core';
-import {TuiExpand} from '@taiga-ui/experimental';
-import {TuiChevron, TuiChip} from '@taiga-ui/kit';
+import {TuiTable} from '@taiga-ui/addon-table';
+import {TuiDay, tuiToInt} from '@taiga-ui/cdk';
+import {TuiButton} from '@taiga-ui/core';
+import {TuiChevron} from '@taiga-ui/kit';
 
-interface Item {
-    firstName: string;
-    lastName: string;
-    role: string;
-    balance: number;
+interface User {
+    readonly dob: TuiDay;
+    readonly name: string;
+    readonly age: number;
+    children?: User[];
+}
+
+const TODAY = TuiDay.currentLocal();
+
+const FIRST = [
+    'John',
+    'Jane',
+    'Jack',
+    'Jill',
+    'James',
+    'Joan',
+    'Jim',
+    'Julia',
+    'Joe',
+    'Julia',
+];
+
+const LAST = [
+    'Smith',
+    'West',
+    'Brown',
+    'Jones',
+    'Davis',
+    'Miller',
+    'Johnson',
+    'Jackson',
+    'Williams',
+    'Wilson',
+];
+
+const DATA: readonly User[] = Array.from({length: 10}, () => {
+    const user = generateUser();
+
+    user.children = Array.from({length: getRandom(0, 5)}, () => generateUser());
+
+    return user;
+});
+
+function getAge(dob: TuiDay): number {
+    const years = TODAY.year - dob.year;
+    const months = TODAY.month - dob.month;
+    const days = TODAY.day - dob.day;
+    const offset = tuiToInt(months > 0 || (!months && days > 9));
+
+    return years + offset;
+}
+
+function generateUser(): User {
+    const dob = TODAY.append({day: -Math.floor(Math.random() * 4000) - 7500}),
+        age = getAge(dob),
+        name = `${LAST[getRandom(0, 9)]}, ${FIRST[getRandom(0, 9)]}`;
+
+    return {name, dob, age};
+}
+
+function getRandom(min: number, max: number): number {
+    return Math.floor(Math.random() * (max - min) + min);
 }
 
 @Component({
     standalone: true,
     imports: [
-        AsyncPipe,
+        FormsModule,
         NgForOf,
+        NgIf,
+        ReactiveFormsModule,
         TuiButton,
         TuiChevron,
-        TuiChip,
-        TuiExpand,
-        TuiFormatNumberPipe,
-        TuiHint,
-        TuiMapperPipe,
         TuiTable,
-        TuiTableExpand,
     ],
     templateUrl: './index.html',
     styleUrls: ['./index.less'],
@@ -36,65 +89,12 @@ interface Item {
     changeDetection,
 })
 export default class Example {
-    protected readonly basicData: Item[] = [
-        {
-            firstName: 'Alex',
-            lastName: 'Inkin',
-            role: 'dev',
-            balance: 1323525,
-        },
-        {
-            firstName: 'Roman',
-            lastName: 'Sedov',
-            role: 'dev',
-            balance: 423242,
-        },
-        {
-            firstName: 'Andrei',
-            lastName: 'Serebrennikov',
-            role: 'dev',
-            balance: 4223242,
-        },
-    ];
+    protected columns = ['children', '#', 'name', 'dob', 'age'];
+    protected readonly data = DATA;
 
-    protected manualOpenData: Item[] = [
-        {
-            firstName: 'Joe',
-            lastName: 'Wilson',
-            role: 'design',
-            balance: 423242,
-        },
-        {
-            firstName: 'Julia',
-            lastName: 'Johnson',
-            role: 'design',
-            balance: 4223242,
-        },
-    ];
+    protected readonly rowState: Record<number, boolean> = {};
 
-    protected readonly customContentData: Item[] = [
-        ...this.basicData,
-        ...this.manualOpenData,
-    ];
-
-    protected readonly columns = ['action', 'firstName', 'lastName', 'role', 'balance'];
-
-    protected manualOpen = false;
-    protected customOpen = false;
-
-    public getSumBalance(people: Item[]): number {
-        return people.reduce((res, item) => {
-            res += item.balance;
-
-            return res;
-        }, 0);
-    }
-
-    protected manualToggle(): void {
-        this.manualOpen = !this.manualOpen;
-    }
-
-    protected customToggle(): void {
-        this.customOpen = !this.customOpen;
+    protected toggleRow(index: number): void {
+        this.rowState[index] = !this.rowState[index];
     }
 }
