+++ projects/demo/src/modules/components/table/index.ts
@@ -20,7 +20,8 @@ export default class Page {
         'Dynamic columns',
         'Footer',
         'Resize a large table',
-        'Expandable',
+        'Expandable rows',
+        'Expandable rows: different ways to toggle rows',
         'Manual sorting',
     ];
 }
